Updated update.py for fh_style_update

+++ geonode/tasks/update.py
@@ -68,43 +68,45 @@ def fh_style_update():
                     password=settings.OGC_SERVER['default']['PASSWORD'])
     gn_style_list = Style.objects.filter(name__icontains='fh').exclude(Q(name__icontains="fhm")|Q(sld_body__icontains='<sld:CssParameter name="fill">#ffff00</sld:CssParameter>'))
     # gn_style_list = Style.objects.filter(name__icontains='fh').exclude(Q(name__icontains="fhm"))
+    gs_style = None
     fh_styles_count = len(gn_style_list)
     ctr = 0
     if gn_style_list is not None:
         fhm_style = cat.get_style("fhm")
         for gn_style in gn_style_list:
             #change style in geoserver
-            print "Style name %s " % gn_style.name
-            gs_style  = cat.get_style(gn_style.name)
-
             try:
                 layer = Layer.objects.get(name=gn_style.name)
-                if layer is not None:
+                gs_style  = cat.get_style(gn_style.name)
+                if gs_style is not None:
+                    gs_style  = cat.get_style(gn_style.name)
                     gs_style.update_body(fhm_style.sld_body)
-                    #change style in geonode
-                    gn_style.sld_body = fhm_style.sld_body
-                    gn_style.save()
-                    #for updating thumbnail
-                    params = {
-                        'layers': layer.typename.encode('utf-8'),
-                        'format': 'image/png8',
-                        'width': 200,
-                        'height': 150,
-                    }
-                    p = "&".join("%s=%s" % item for item in params.items())
-                    thumbnail_remote_url = ogc_server_settings.PUBLIC_LOCATION + \
-                        "wms/reflect?" + p
-                    # thumbnail_create_url = ogc_server_settings.LOCATION + \
-                    #     "wms/reflect?" + p
-                    create_thumbnail(layer, thumbnail_remote_url, thumbnail_remote_url, ogc_client=http_client)
-                    ctr+=1
-                    print "'{0}' out of '{1}' : Updated style for '{2}' ".format(ctr,fh_styles_count,gn_style.name)
+                else:
+                    cat.create_style(gn_style.name,fhm.sld_body)
+                #change style in geonode
+                gn_style.sld_body = fhm_style.sld_body
+                gn_style.save()
+                #for updating thumbnail
+                params = {
+                    'layers': layer.typename.encode('utf-8'),
+                    'format': 'image/png8',
+                    'width': 200,
+                    'height': 150,
+                }
+                p = "&".join("%s=%s" % item for item in params.items())
+                thumbnail_remote_url = ogc_server_settings.PUBLIC_LOCATION + \
+                    "wms/reflect?" + p
+                # thumbnail_create_url = ogc_server_settings.LOCATION + \
+                #     "wms/reflect?" + p
+                create_thumbnail(layer, thumbnail_remote_url, thumbnail_remote_url, ogc_client=http_client)
+                ctr+=1
+                print "'{0}' out of '{1}' : Updated style for '{2}' ".format(ctr,fh_styles_count,gn_style.name)
+            except Exception as e:
+                err_msg = str(e)
+                if "Layer matching query does not exist" in err_msg:
+                    gn_style.delete()
 
-            except:
-            #if a style with no layer exists
-                print "Will delete %s" % gn_style.name
-                gn_style.delete()
-                cat.delete(gs_style)
+
 
 @task(name='geonode.tasks.update.ceph_metadata_udate', queue='update')
 def ceph_metadata_udate(uploaded_objects):
